requirements.txt
# ============================================
# FISCALIA - Dependências do Projeto
# ============================================

# Core Framework
crewai>=0.140.0
crewai-tools>=0.15.0
langchain==0.3.12
langchain-groq==0.2.1
langchain-openai==0.2.14

# LLM Providers
groq>=0.11.0
openai>=1.54.3

# Web Interface
streamlit>=1.46.1
streamlit-option-menu>=0.4.0

# Database
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.10
alembic>=1.13.3

# Data Processing
pandas>=2.2.3
numpy>=2.1.3
openpyxl>=3.1.5

# XML Processing
lxml==5.3.0
xmltodict==0.14.2

# File Handling
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'

# Validation
pydantic==2.10.6
pydantic-settings==2.7.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pytz==2024.2

# Logging
loguru==0.7.2

# Testing
pytest==8.3.3
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-asyncio==0.21.1

# Development Tools
black==24.10.0
flake8==7.1.1
mypy==1.13.0
pre-commit==4.0.1

# Railway/Docker
gunicorn==23.0.0
uvicorn==0.32.0
